Fixed the negative numbers on syscalls issue.  Thanks to Bander!

+++ include/utils/types.h
@@ -4,7 +4,7 @@
 
 #ifndef HIGHPERFORMANCELINUXSENSORS_TYPES_H
 #define HIGHPERFORMANCELINUXSENSORS_TYPES_H
-typedef short pid_t;
-typedef short syscall_t;
+typedef int pid_t;
+typedef int syscall_t;
 //typedef int syscall_t;
 #endif //HIGHPERFORMANCELINUXSENSORS_TYPES_H
